+++ package/android/src/main/java/com/mrousavy/camera/core/CodeScannerPipeline.kt
@@ -7,7 +7,6 @@ import com.google.mlkit.vision.barcode.BarcodeScanner
 import com.google.mlkit.vision.barcode.BarcodeScannerOptions
 import com.google.mlkit.vision.barcode.BarcodeScanning
 import com.google.mlkit.vision.common.InputImage
-import com.mrousavy.camera.CameraQueues
 import com.mrousavy.camera.core.outputs.CameraOutputs
 import com.mrousavy.camera.parsers.Orientation
 import java.io.Closeable
