+++ package/android/src/main/java/com/mrousavy/camera/extensions/CameraCaptureSession+capture.kt
@@ -5,9 +5,9 @@ import android.hardware.camera2.CaptureFailure
 import android.hardware.camera2.CaptureRequest
 import android.hardware.camera2.TotalCaptureResult
 import android.media.MediaActionSound
-import com.mrousavy.camera.CameraQueues
-import com.mrousavy.camera.CaptureAbortedError
-import com.mrousavy.camera.UnknownCaptureError
+import com.mrousavy.camera.core.CameraQueues
+import com.mrousavy.camera.core.CaptureAbortedError
+import com.mrousavy.camera.core.UnknownCaptureError
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
 import kotlin.coroutines.suspendCoroutine
