+++ package/android/src/main/java/com/mrousavy/camera/extensions/CameraDevice+createCaptureSession.kt
@@ -8,8 +8,8 @@ import android.hardware.camera2.params.OutputConfiguration
 import android.hardware.camera2.params.SessionConfiguration
 import android.os.Build
 import android.util.Log
-import com.mrousavy.camera.CameraQueues
-import com.mrousavy.camera.CameraSessionCannotBeConfiguredError
+import com.mrousavy.camera.core.CameraQueues
+import com.mrousavy.camera.core.CameraSessionCannotBeConfiguredError
 import com.mrousavy.camera.core.outputs.CameraOutputs
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
