+++ package/android/src/main/java/com/mrousavy/camera/extensions/CameraManager+openCamera.kt
@@ -5,9 +5,9 @@ import android.hardware.camera2.CameraDevice
 import android.hardware.camera2.CameraManager
 import android.os.Build
 import android.util.Log
-import com.mrousavy.camera.CameraCannotBeOpenedError
-import com.mrousavy.camera.CameraDisconnectedError
-import com.mrousavy.camera.CameraQueues
+import com.mrousavy.camera.core.CameraCannotBeOpenedError
+import com.mrousavy.camera.core.CameraDisconnectedError
+import com.mrousavy.camera.core.CameraQueues
 import com.mrousavy.camera.parsers.CameraDeviceError
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
