+++ package/android/src/main/java/com/mrousavy/camera/frameprocessor/Frame.java
@@ -4,11 +4,9 @@ import android.hardware.HardwareBuffer;
 import android.media.Image;
 import android.os.Build;
 import com.facebook.proguard.annotations.DoNotStrip;
-import com.mrousavy.camera.HardwareBuffersNotAvailableError;
+import com.mrousavy.camera.core.HardwareBuffersNotAvailableError;
 import com.mrousavy.camera.parsers.PixelFormat;
 import com.mrousavy.camera.parsers.Orientation;
-
-import java.nio.ByteBuffer;
 
 public class Frame {
     private final Image image;
