+++ package/ios/Core/CameraSession.swift
@@ -1,5 +1,5 @@
 //
-//  CaptureSession.swift
+//  CameraSession.swift
 //  VisionCamera
 //
 //  Created by Marc Rousavy on 11.10.23.
@@ -9,6 +9,6 @@
 import Foundation
 import AVFoundation
 
-class CaptureSession {
+class CameraSession {
   
 }
